refactor: consolidate DTO type alias files into single module (#326)
Consolidate 8 single-line type alias files into single_task_inputs.py
to reduce file proliferation and improve maintainability.

- Create single_task_inputs.py with all type aliases
- Update __init__.py to export from new module
- Update imports in use cases, controllers, and tests
- Delete 8 individual type alias files
