--- packages/taskdog-core/src/taskdog_core/application/dto/reopen_task_input.py
@@ -1,6 +1,0 @@
-"""DTO for reopening a task."""
-
-from taskdog_core.application.dto.base import SingleTaskInput
-
-# Type alias for clarity
-ReopenTaskInput = SingleTaskInput
