--- packages/taskdog-core/src/taskdog_core/application/dto/restore_task_input.py
@@ -1,6 +1,0 @@
-"""DTO for restoring a task."""
-
-from taskdog_core.application.dto.base import SingleTaskInput
-
-# Type alias for semantic clarity
-RestoreTaskInput = SingleTaskInput
